pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pipclaw"
version = "0.1.0"
description = "CLI-based AI agent powered by Kimi K2.5"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0",
    "textual>=0.60",
    "rich>=13.0",
    "python-dotenv>=1.0",
    "pyperclip>=1.8",
    "pathspec>=0.12",
    "term-image>=0.7",
    "mutagen>=1.47",
]

[project.optional-dependencies]
pdf = ["pymupdf>=1.24"]

[project.scripts]
pipclaw = "pipclaw.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/pipclaw"]
